+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
-## 1.0.0
+## 1.1.2
 
-Initial release.
+Switch to YAML scans for `p3rpc.flowscriptframework`
+
+## 1.1.0/1.1.1
+
+Added `IScriptState` method to invoke the original Flowscript function